+++ tubelib_addons1/harvester.lua
@@ -414,6 +414,10 @@ minetest.register_craft({
 
 
 tubelib.register_node("tubelib_addons1:harvester_base", {}, {
+	on_pull_stack = function(pos, side)
+		local meta = minetest.get_meta(pos)
+		return tubelib.get_stack(meta, "main")
+	end,
 	on_pull_item = function(pos, side)
 		local meta = minetest.get_meta(pos)
 		return tubelib.get_item(meta, "main")